Initial commit - FoodNDeliv API Security by Design

+++ src/main/java/com/example/foodndeliv/dto/OrderLineDTO.java
@@ -2,18 +2,19 @@ package com.example.foodndeliv.dto;
 
 import lombok.Data;
 import jakarta.validation.constraints.*;
+import java.math.BigDecimal;
 
 @Data
 public class OrderLineDTO {
-    
+
     @NotBlank(message = "Product name must not be blank")
     private String productName;
-    
+
     @NotNull(message = "Quantity must not be null")
     @Positive(message = "Quantity must be positive")
     private Integer quantity;
-    
+
     @NotNull(message = "Price must not be null")
     @DecimalMin(value = "0.01", message = "Price must be at least 0.01")
-    private Double price;
+    private BigDecimal price;
 }
